backend/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Limiter==3.12
psycopg2-binary==2.9.10
python-dotenv==1.1.1
PyJWT==2.10.1
gunicorn==21.2.0
marshmallow==3.20.1
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0